pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "flask-socketio>=5.5.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "flask-limiter>=3.12",
    "bleach>=6.2.0",
    "streamlit>=1.45.1",
    "streamlit-chat>=0.1.1",
    "streamlit-authenticator>=0.4.2",
    "streamlit-option-menu>=0.4.0",
    "websockets>=15.0.1",
    "asyncio>=3.4.3",
    "pandas>=2.3.0",
    "gevent>=25.5.1",
    "gevent-websocket>=0.10.1",
    "flask-wtf>=1.2.2",
]
